move on 32b instead of 64b

+++ assembly/__tests__/fast/capture-ordering.spec.ts
@@ -67,7 +67,7 @@ describe('"Most valuable victim, least valuable attacker" capture comparison', (
       QUEEN + BLACK,
       17
     );
-    const captures: u64[] = [
+    const captures: u32[] = [
       lowValueCapture,
       highestValueCapture,
       mediumValueCapture,
