+++ assembly/__tests__/fast/sliding-pieces-captures.spec.ts
@@ -21,7 +21,7 @@ describe("Rook captures", () => {
     board.putPiece(ROOK, WHITE, 0);
     board.putPiece(KNIGHT, WHITE, 3);
     // when
-    const captures: u64[] = [];
+    const captures: u32[] = [];
     addRookPseudoLegalCaptures(captures, board, WHITE);
     // then
     expect(captures).toHaveLength(1);
@@ -37,7 +37,7 @@ describe("Bishop captures", () => {
     board.putPiece(BISHOP, WHITE, 36);
     board.putPiece(KNIGHT, WHITE, 43);
     // when
-    const captures: u64[] = [];
+    const captures: u32[] = [];
     addBishopPseudoLegalCaptures(captures, board, WHITE);
     // then
     expect(captures).toHaveLength(1);
@@ -53,7 +53,7 @@ describe("Queen captures", () => {
     board.putPiece(QUEEN, WHITE, 36);
     board.putPiece(KNIGHT, WHITE, 43);
     // when
-    const captures: u64[] = [];
+    const captures: u32[] = [];
     addQueenPseudoLegalCaptures(captures, board, WHITE);
     // then
     expect(captures).toHaveLength(1);
